Refactor - Add a new component Results  🔪
